+++ packages/core/src/directives/reports/createCommentDirectiveNoSelection.ts
@@ -1,9 +1,12 @@
-import type { NormalizedReportRangeObject } from "../../types/reports.ts";
+import type {
+	FileReport,
+	NormalizedReportRangeObject,
+} from "../../types/reports.ts";
 
 export function createCommentDirectiveNoSelection(
 	type: string,
 	range: NormalizedReportRangeObject,
-) {
+): FileReport {
 	return {
 		about: {
 			id: "commentDirectiveNoSelection",