+++ packages/core/src/types/host.ts
@@ -5,6 +5,7 @@ export interface LinterHost {
 	getCurrentDirectory(): string;
 	getFileTouchTime(filePath: string): Promise<number | undefined>;
 	getFileTouchTimeSync(filePath: string): number | undefined;
+	getRepositoryRoot(): string | undefined;
 
 	/**
 	 * Find a set of files relative to the cwd given a set of glob patterns.