+++ packages/ts/src/rules/typeImports.ts
@@ -312,7 +312,7 @@ export default ruleCreator.createRule(typescriptLanguage, {
 		description:
 			"Reports imports that do not match the configured type import style.",
 		id: "typeImports",
-		presets: ["stylistic"],
+		presets: ["logicalStrict"],
 	},
 	messages: {
 		avoidImportType: {