+++ packages/ts/src/rules/utils/getRegExpLiteralDetails.ts
@@ -3,10 +3,16 @@ import type {
 	TypeScriptFileServices,
 } from "@flint.fyi/typescript-language";
 
+export interface RegExpLiteralDetails {
+	flags: string;
+	pattern: string;
+	start: number;
+}
+
 export function getRegExpLiteralDetails(
 	node: AST.RegularExpressionLiteral,
 	{ sourceFile }: Pick<TypeScriptFileServices, "sourceFile">,
-) {
+): RegExpLiteralDetails {
 	const lastSlash = node.text.lastIndexOf("/");
 	return {
 		flags: node.text.slice(lastSlash + 1),