+++ packages/typescript-language/src/utils/declarationIncludesGlobal.ts
@@ -3,7 +3,7 @@ import type { Declaration, Program } from "typescript";
 export function declarationIncludesGlobal(
 	declaration: Declaration,
 	program: Program,
-) {
+): boolean {
 	const sourceFile = declaration.getSourceFile();
 	return (
 		program.isSourceFileDefaultLibrary(sourceFile) ||